introcude common interface for gpu graphs
